Move max heap into separate subpackage
